feat(library): add reusable edit button for images
This commit introduces a new "Edit" button to the image details dialog in the library, allowing users to send an image directly to the Gemini Image Generation page for editing.

Key changes include:
- Created a new reusable `edit_button` component to encapsulate the navigation logic.
- Modified the Gemini Image Generation page to accept an `image_uri` query parameter on load, pre-populating the editor.
- Added the new `edit_button` to the library's image detail view.

refactor(navigation): standardize on 'image_uri' query parameter

To improve consistency across the application, the query parameter for passing image GCS URIs between pages has been standardized to `image_uri`.

- Refactored the Veo page (`pages/veo.py`) to expect `image_uri` instead of `source_image_uri`.
- Updated the "Send to Veo" buttons in the library and on the Gemini Image Generation page to send the new standardized parameter.

chore(config): implement environment-specific configurations

Implements a system for loading environment-specific configurations, particularly for UI content.

- The application now loads `config/about_content.<APP_ENV>.json` if it exists, falling back to the default `config/about_content.json`.
- Added a `.gcloudignore` file to ensure these environment-specific files are included in Cloud Run deployments.
- Updated `README.md` to document the new configuration feature.

refactor(buttons): extract Veo button into a reusable component

Following the pattern established with the edit_button, the "Send to Veo" button has been extracted into its own reusable component.

- Created a new `veo_button` component that encapsulates navigation logic.
- Refactored the library details page and the Gemini Image Generation page to use the new, self-contained component. This improves modularity and simplifies adding this action to other pages.
